fcoeadm: Integrating fcoeadm management tool with HBAAPI
Previously the fcoeadm management tool does not invoke the HBAAPI
library. This patch re-implements the options of the fcoeadm
management tool to invoke the HBAAPI library, i.e. libHBAAPI.so,
which indirectly invokes the vendor library libhbalinux.so. More
information of the FCoE interfaces, local ports, remote ports,
and LUNs may be displayed. For instrunctions of how to build and
install the tool, please see usr/INSTALL file. For the command
line options please the man page in the usr/doc directory.

The fcoeadm management tool requires that libHBAAPI and libhbalinuz
are both installed on the Linux system. The source code and
instructions can be found on www.Open-FCoE.org.

Signed-off-by: Steve Ma <steve.ma@intel.com>