fcoeadm: Correct the install directory and improve Makefile
The install directory of fcoeadm and the future tools are to be
installed into /sbin. This patch fixes the Makefile to install
tools in the required directory and allow installing multiple
tool programs in a loop. This patch also allow the Makefile
to generate a global version.c file for all the tools.

Signed-off-by: Steve Ma <steve.ma@intel.com>