fcoeadm: Fix reset defect with large host_no
Currently we create a 64 byte string for the fc_host directory name.
64 bytes should be more than enough, but when we pass forward the
length we pass the sizeof(fchost) which is the size of the char *,
not the lenght of the string. This causes problems if the host_no
becomes large. For example on x86_64 we will pass a length of 8
forward and when we try and copy the directory name "host2214" we
were copying (len - 1) which means that the fchost buffer would only
be filled with "host221" and when fcoeadm attempted to access the
directory it did not exist.

This fix simply passes the length of the buffer forward into the
validation and buffer copying functions.

This patch also uses strnlen to determine the size of the directory
name. I had to compile with '-D _GNU_SOURCE' to use strnlen.

Signed-off-by: Robert Love <robert.w.love@intel.com>