open-fcoe: Initial code for fcoemon and Open-FCoE service script
fcoemon is the daemon of the fcoe system service to interact with
the dcbd daemon (Data Center Bridging daemon) and the kernel
RT_NETLINK service. When fcoemon starts, it establishes a socket
connection with the dcbd daemon. It then sends commands to, and
receives responses and event from the DCB daemon.  The RT_NETLINK
service reports the link status of the Ethernet ports to the fcoemon.
The fcoemon will process the responses and events received from
dcbd and invoke the fcoeplumb script to add multiq queue descipline,
to add/remove/replace the skbedit filters, and create/destroy the
FCoE interfaces.

The "Open-FCoE service script" is the bash script, /etc/init.d.fcoe,
after the Open-FCoE user-space package is installed. This script,
just like the other system service scripts, starts, stops, restarts,
and shows status of the fcoe service. The script will create FCoE
interface for an Ethernet port directly using fcoeadm tool if the
port is not configured to use DCB service. Otherwise, if the port
is configured to use DCB service, it will start the fcoemon daemon,
and let the fcoemon to add multiq qdiscipline and skbedit filter.
The man page of fcoemon describes more in this area.

The configuration files include
    /etc/fcoe/config
    /etc/fcoe/cfg-eth0, ..., /etc/fcoe/cfg-eth<n> one for each
    Ethernet port.
These files are self-documented, and must be edited to contain the
settings for your environment.

Notice that the fcoe system service does not depend on dcbd service.
However, the fcoemon daemon will be started by the fcoe service only
if any one of the Ethernet port requires DCB service. In this case,
fcoe service depends on dcbd service. If none of the Ethernet ports
requires DCB service, the fcoemon will not be started and, in this
case, the fcoe service does not depend on dcbd service.

With this patch applied on the Open-FCoE user-space source tree, the
installation should contain the following files:

    /etc/init.d/fcoe                  startup script
    /etc/fcoe/config                  Common configuration file
                                       (currently only used to specify
                                       debugg options)
    /etc/fcoe/cfg-eth*                Per-port configuration files
    /etc/fcoe/scripts/fcoeplumb       Script only invoked by fcoemon
                                       to add qdisc, filters and
                                       create/destroy FCoE interfaces.
    /sbin/fcoeadm                     FCoE management tool
    /sbin/fcoemon                     FCoE monitor
    /usr/share/man/man8/fcoeadm.8.gz  Man-page for fcoeadm
    /usr/share/man/man8/fcoemon.8.gz  Man-page for fcoemon

For installation, please see the INSTALL file included in the patch.

The Open-FCoE service script has not been well-developed, and heavily
tested. More patches are expected to be available in the near future.
The DCB package is available to be downloaded from the sourceForge web site.

Signed-off-by: Steve Ma <steve.ma@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>