fcoemon: fcoemon doesn't destroy fcoe interface when switch exit fcoe mode
When switch leaves fcoe mode 'no fcoe mode' fcoemon does not detect that app
config operational state is false 'dcbtool go eth2 app:0'.  In this case I
would expect that fcoemon removes the interface it does not.

This patch also fixed a few logging messages and inhibits the pinging messages
to the dcbd.

Signed-off-by: Steve Ma <steve.ma@intel.com>
Signed-off-by: Chris Leech <christopher.leech@intel.com>