fcoemon: fcoemon doesn't validate DCB state and issue error messages
The fcoemon requires the DCB state is on for creating fcoe interface.
Previously "dcbtool sc ethx dcb on" is issued in the fcoe service
init script to insure the DCB state is on. Since the fcoe init script
is no longer issuing dcbtool command in order to allow dcbd service
to start after the fcoe service, we need to add code to validate the
DCB state and log errors if it is off. Because configuring APP:0 will
fail if the DCB state is off, checking the DCB state must occur before
setting the DCB APP:0 configuration.  This patch also added error log
messages while validating the other DCB parameters. Users are
responsible to turn on DCB state of an Ethernet port for the system.
This is a one-time operation over the reboots unless the Ethernet
configuration is altered or DCB configuration is reset.

Also fixed a few coding style errors introduced by previous code
drops.

Signed-off-by: Steve Ma <steve.ma@intel.com>
Signed-off-by: Chris Leech <christopher.leech@intel.com>