fcoe-utils: add --without-dcb mode for backports
I've been asked to provide a means of building the fcoe-utils without DCB
support, for distros that backport FCoE but not DCB in their kernel.

This just skips fcoemon and it's man page if you specify --without-dcb,
and the dcbd development package is not required.

It works if you set DCB_REQUIRED="no" in the configuration files.

Signed-off-by: Chris Leech <christopher.leech@intel.com>