fcoeplumb: setup tc filter for ETH_P_FCOE, not ETH_P_802_3
The FCoE changes from 2.6.29 to 2.6.30 include chaning the
skb->protocol field from ETH_P_802_3 to ETH_P_FCOE.  This impacts the
tc filters used to map FCoE frames to the correct traffic class for DCB.

This patch changes the tc filter setup in the fcoeplumb script to match.
This is needed for 2.6.30 FCoE use with DCB, but is not backwards
compatible with the 2.6.29 kernel code.

Signed-off-by: Robert Love <robert.w.love@intel.com>