fcoe-utils: Add debug/ directory and dcbcheck.sh script
Adds a debug/ directory in the source tree intended to contain
debugging tools.

The first debugging tool, that this patch also adds, is
dcbcheck.sh. This script will use 'dcbtool' to check if
DCB is configured correctly for FCoE traffic. If some
criteria is not met the script will suggest commands to
run to get to the user to a correctly configured state.

Signed-off-by: Robert Love <robert.w.love@intel.com>