fcoeplumb: Add filter for FIP ethertype
FIP code has been added to the kernel and we want the
QoS layer in the kernel to filter FIP traffic into the
DCB negotiated traffic class. This patch addd an extra
'tc' call to fcoeplumb to filter the FIP packets.

This patch makes many of the filter functions generic
to work with multiple filters (one per ethertype) instead
of assuming that FCoE is the only desired filter.

This patch also fixes three defects-

1) incorrect argument parsing

 delete_skbedit_filter()
 {
        ifname=$1
-       queue=$?
+       queue=$2

$? doesn't make sense here

2) incorrect awk regex

The first check in the find_skbedit_filter() was searching for
/^filter.*parent.*protocol '$ethertype' ....

However, there are squared brackets around the ethertype. Since
the regex was searching for "protocol"<space><ethtype> it was
not matching "protocol"<space>[<ethertype>

This patch adds the brackets to the check.

3) When no qdisc was found the script would add a qdisc
   and filter, but then immediately delete the newly added
   qdisc. It would then re-add it and the filters again. This
   was unnecessary, so I removed it.

   It was explained to me that this was a workaround for a
   'tc' defect, but with the latest version of iproute2
   this workaround is not needed.