fcoe-utils: fcoeadm -i returns confusing message when interface is not created
If a user executed the following command "fcoeadm -i ethX" when ethX is not
currently created as an fcoe device the following message will be returned:

# fcoeadm -i eth2
fcoeadm: No fc_host found for eth2
fcoeadm: Invalid command options!
Usage: fcoeadm
         [-c|--create] <ethX>
         [-d|--destroy] <ethX>
         [-r|--reset] <ethX>
         [-i|--interface] [<ethX>]
         [-t|--target] [<ethX>]
         [-l|--lun] [<target port_id> [<lun_id>]]
         [-s|--stats] <ethX> [-n <interval>]
         [-h|--help]

The user may be confused by the inclusion of the text "invalid command options!".
If the output is "No fc_host found for eth2". It is sufficient and would
not confuse a user.

Signed-off-by: Steve Ma <steve.ma@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>