fcoe-utils: Reduce debug flags to one
There are currently two debug flags that are checked
through the code. fcoemon isn't complicated enough to require
this level of granularity regarding developer debug statements.

Also, these developer debug statements were enabled by default
even though there is an option for the program to enable them.

This patch reduces the number of debug flags to one and only
sets it when the '-d' or '--debug' option(s) are supplied by
the user.

Signed-off-by: Robert Love <robert.w.love@intel.com>