fcoe-utils: fcoeplumb takes "--debug" argument from caller, not config file
Currently fcoemon is sourcing the /etc/fcoe/config file to determine if
it should be running in debug mode. Generally this script will be called
from fcoemon, which was called from /etc/init.d/fcoe. Currently, all
three of these entities are reading the config file. Eventually, only
the init script will read the config file and the debug flag will be
passed from one entity to the other.

A first step towards this goal is to not have fcoeplumb read "DEBUG"
from the config file and instead take that option as an argument to
the script.

Signed-off-by: Robert Love <robert.w.love@intel.com>