fcoe-utils: Don't depend on "eth" as part of the interface name
Currently fcoemon will not parse a config file for an interface
if it does not start with "cfg-eth". Depending on "cfg-" makes
sense becuase these are the fcoemon config files, but we don't
want to restrict the names of the interfaces that fcoemon will
manage.

This patch makes it so that fcoemon will manage interfaces with
any name as long as their config file starts with "cfg-".

Signed-off-by: Robert Love <robert.w.love@intel.com>