fcoe-utils: Ignore interface operational state
Currently fcoemon has a set of criteria that it uses to determine
if an interface should be created or destroyed. The current set
includes the link state (i.e. operational state). What this means
is that fcoemon will not try to create the interface if there is
no link. It also means that if the link goes down on an already
created interface then it will destroy that interface.

The libfc/fcoe protocol stack in the kernel already handles link
events. Generally, on link down the stack blocks I/O and on link
up it un-blocks I/O. So, if the Ethernet cable is pulled out, I/O
will pause. If that same cable is re-inserted then the I/O will
resume.

The problem is that fcoemon is interfering with the kernel's
ability to resume I/O. This is happening becuase fcoemon will
receive the link down event. It will see that it does not meet
it's criteria and it will destory the interface. This destory
frees all of the kernel's per-instance data for that interface
and when the link becomes active again the I/O cannot be restarted
becuase the required structures have been freed.

This patch removes the 'operational state' criteria from the
set of requirements fcoemon uses when determining whether to
create or destroy an interface.

Interesting notes:

This patch does not modify the link notification code at all and
we are still caching the link state.

Also, link down events will generate a few dcbd operational change
event. None of these events alters the behavior of fcoemon (regarding
create/destroy of interfaces) so this should be fine.

Signed-off-by: Robert Love <robert.w.love@intel.com>