fcoe-utils: fcping implementation via FC pass-through interface
This is the implementation of fcping by invoking the FCoE ELS/CT
pass-through interface in the kernel. fcping is added as one of the
command tools in fcoe-utils. Pinging is to send an ECHO ELS from
a local port of an HBA to a remote FCoE destination(e.g. target,
directory server, etc.). Issue fcping -h to see a summary of the
command usage.  libHBAAPI and libhbalinux libraries are used by
fcping to find HBA information.

VLAN Handling:

Because the VLAN interface can have any name, we cannot
just check the name to see if it is "ethxxx". The patch
is to check in /sys/class/net to see if the interface
name (VLAN or non-VLAN) exists in this directory as a
symbolic link. If yes, then pass the name to libhbalinux
to check if it is an FCoE interface. If no, proceed to
check if it is other form of representation of the
interface (VLAN or non-VLAN).

Also updated the INSTALL document and the QUICKSTART document
for fcping.

Signed-off-by: Steve Ma <steve.ma@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>