fcoe-utils: Change sa_select_loop to have a return value
It is desired to move the debug macros out of fcoemon_debug.h
and into fcoemon.h. Since fcoemon_utils is a utility library
that could be used for other applications it can provide a
logging facility, but the specific macros for each application
should be in each applications header files.

This means that any of the fcoemon_utils.[ch] helper functions
should not do any logging themselves, becuase they are part of
the utility library and not the application.

sa_select_loop() is the only example of a helper function that
does logging too. It logs when select() fails.

This patch has sa_select_loop() return 1 when it fails so the
caller can log the error message and exit. It will return 0
on success.

Signed-off-by: Robert Love <robert.w.love@intel.com>