fcoe-utils: Add "--syslog" flag to fcoeplumb
fcoeplumb should not re-read the /etc/fcoe/config file.
Instead have this option and allow the caller (fceomon)
to pass it.

This patch also introduces a config_logging routine that
will be used when either the --debug or --syslog flags are
passed to the script. This makes the order of these flags
irrelevant as it will be called when either flag is received.

Also, modify fcoemon to pass the --syslog flag if use_syslog
is set.

Signed-off-by: Robert Love <robert.w.love@intel.com>