fcoe-utils: Fix incorrect statements in fcoemon's man page
This patch corrects some of the language in the fcoemon man
page to make it more correct.

It also, converts all references to the 'fcoe' service, fcoeadm
or fcoemon to be use the BOLD font property. Previously it was
inconsistent, some instances were underlined, some were BOLD
and some had no modification. Lastly, this patch ensures that
any paths are underlined.

Signed-off-by: Robert Love <robert.w.love@intel.com>