fcoe-utils: add fcc.sh config script
I find this script handy in addition to fcoeadm.
It shows HBA status, statistics, and remote port and LUN status.
It can do create/delete/reset operations as well.
However, it does not interact with fcoemon.

Other than create/delete, it shows all kinds of HBAs, even those
that don't use libfc, and it doesn't require HBA-API libraries
or even compilation, so it is easy to install.

Signed-off-by: Joe Eykholt <jeykholt@cisco.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>