fcoe-utils: fcoeplumb, add --qos-enable and --qos-disable commands
This changes fcoeplumb so that it will only remove or add
filters when the "--qos-enable" or "--qos-disable" option is
specified.  This will be useful for devices that do not need
to monitor or configure QOS settings.

This requires fcoemon to supply the "--qos-enable" or
"--qos-disable"  flag to fcoeplumb when it disables an interface.

Additionally --enable, --disable are renamed --create and
--destroy to be inline with sysfs and their functionality.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>