fcoe-utils: only add interfaces to the ff_list that have config files
Currently, all interfaces are added to the ff_list regradless of if they
will be used or not.  Additionally, they are added using the ifindex
number.  This requires checks to be added to verify that the interfaces in
ff_list are valid before using them.  Also if the ifindex number of an
interface changes for instance if the netdevice module is unloaded
and reloaded fcoemon will no longer correctly monitor that interface.

This patch only adds interfaces to the ff_list that have valid
configuration files.  Also the lookup is done by ifname to protect
against the scenario above.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>