fcoe-utils: add support to query vlan devices in fcoemon.
This adds support to process the IFLA_LINKINFO in the nlmsg.
to determine if a device is a vlan. Also, functions are added
to parse /proc/net/vlan/ for the virtual interfaces real
device name.  This information is not provided through rtnetlink.

If the device is a vlan do not monitor it.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>