fcoe-utils: fcoemon: fcm_fcoe_lookup_name does not return NULL correctly
fcm_fcoe_lookup_name() searches the ff_list for matching fcm_fcoe
interfaces.  When this function is used the return value is checked
to see if it is NULL to indicate that a match was not found.

This fixes fcm_fcoe_lookup_name to return NULL when no match is found.
Previously, it would return the last element in the list.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>