fcoe-utils: fcping: fix use of non-ethernet hba specs
Since /sys/class/net/ethX is not a symlink, readlink would always fail,
but that was erroneously taken as an indication that the interface
was present!  So, none of the other ways of specifying an hba would
be tried.

For the fnic driver, we need to specify the hba as hostX or by WWN.

Use stat(2) instead of readlink(2) to determine whether
/sys/class/net/ethX is present.

Signed-off-by: Joe Eykholt <jeykholt@cisco.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>