fcoe-utils: fcping: fix ability to specify target by WWPN
The ability to specify the target of fcping with WWPN or WWNN
didn't work because fp_ns_get_id() wasn't using the bsg SG_IO
interface correctly.

The destination port_id and the preamble of the CT payload needs
to be filled in in the request.  Also, the response comes back as
a struct fc_bsg_reply.  The opcode was specified in host order.

Signed-off-by: Joe Eykholt <jeykholt@cisco.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>