fcoe-utils: fix fcoemon to also set app config data for vlans
This fixes a bug introduced with vlan support where fcoemon will
disable the app feature after some error cases.

fcoemon clears the app config data when it goes
into an error state.  If fcoemon leaves the error state it
queries the configuration data to see if the app feature
should be enabled and then enables it if this is true.
The search for this was not searching the vlan configurations.
This resulted in the app config data not being set correctly.
Also because this was used to set the app feature it caused
fcoemon to disable the app feature completly.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>