fcoe-utils: do not monitor dcbd event when interface is down
Add a check for ifi->flags to only accept dcbd events when the
lower interface bit is set (IFF_LOWER_UP). There is no need to
monitor events when the interface is down the stack will handle
this case.  This removes some unecessary create/destroy sequences.

Also when an fcoe interface is created the interface may need to be
reset to enable the hardware. This reset is causing the dcbd operstate
to become false in turn fcoemon destroys the fcoe interface.
When the interface comes back online the operstate becomes true
again and fcoemon will create the fcoe interface.  This loop can
continue indefinantly.

This patch also changes the state to FCD_INIT when the device
is enabled and FCD_ERROR when the device is disabled. This stops
the state machine while the interface is down then restarts it
when the link is brought back up.  This is needed to account for
real dcb changes occuring while the interface is down.

Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>