fcoe-utils: fcrls: add read link status tool (RLS)
Sends the ELS_RLS (Read Link Status) command to the target rport to retrieve
the Link Error Status Block (LESB) as described by the FC-BB-5 Rev2. The
target rport is assumed to be already PLOGI-ed from where this command is
issued.

This tool is based on existing fcping/fcnsq tool, however as ELS_RLS must
be sent to a logged-in port, fcrls sends this command via FC_BSG_RPT_ELS,
not FC_BSG_HST_ELS_NOLOGIN.

Signed-off-by: Yi Zou <yi.zou@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>