Obtain the real device of a VLAN interface using an ioctl.
Obtaining the real device of a VLAN interface by parsing /proc was sometimes
failing due to fcoemon attempting to read the /proc vlan entry before it was
created.  This patch uses the GET_VLAN_REALDEV_NAME_CMD ioctl to get the real
device name.  This results in simpler and less code.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>