Fixup dcbd connection timeout code
The connection to dcbd was not getting correctly re-established when
dcbd was stopped and restarted quickly.   Sometimes the connection to
dcbd was established without registering to receive dcbd events - which
resulted in crippled fcoemon functionality from that point on.
This patch fixes the dcbd connection logic so that it always retried
in a timely manner and that registering for dcbd events always occurs.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>