Add separate arguments to specify FCoE and network interface
The network interface upon which the FCoE interface is created may be

different than the network interface upon which the qdisc and tc filters
are configured.  This patch adds argument flags to identify each interface
and is correlated with the appropriate action arguments (i.e. create/destroy
command or qos commands.)

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>