Update the FCoE and network interface lists.
fcoemon has a number of problems related to properly creating or destroying
FCoE intefaces  due to how its state machine is structured and managed.  A
representave example of one kind of problem is:
- if the FCoE interface is configured on a VLAN, removing the VLAN and then
  re-adding the VLAN will result in the FCoE interface being destroyed (OK),
  but not getting recreated (not OK).

This issue stem from the fact that fcoemon manages the FCoE interface state
via a single data structure - a list of network interfaces.  The FCoE
interface state is maintained in a network interface element.  This is not
quite correct since a single network interface might support multiple FCoE
interfaces on different VLANs.

The new approach provided by this patch repurposes the fcoemon code and
structures to maintain two seperate lists - each with a clear purpose.
First, there is an FCoE interface list which is initialized from the FCoE
interface configuration files.  This list maintains the the FCoE interface
configuration, which network interface it uses, the (possibly different)
real network interface it uses (learned dynamically), and it tracks the FCoE
interface actions (i.e. create, destroy) which must occur.  The second list
is a list of real network interfaces which are used by one or more of the
FCoE interfaces.  The network interface list tracks the operstate of the
network interface and it tracks the dcbd query status of the network
interface if one or more the FCoE interfaces which use it require DCB.

There are two other changes to the network interface list.  1) only network
interfaces needed for FCoE interfaces are maintained in the list - instead
of all detected network interfaces.  2) A sub VLAN list is not maintained.
The only VLAN interfaces that are of interest are those which are associated
with an FCoE interface.  Thus the relevant VLAN interface information is
already present in the FCoE interface list - no need to duplicate it under
the network interface list.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>