Update the fcoemon state machine to be insensitive to intermittent dcb changes
fcoemon has a number of problems related to properly creating or destroying
FCoE intefaces  due to how its state machine is structured.  A
representative issue is:
- a temporary link down (due to an arbitrary network interface configuration
  or whatever) will cause dcbd to re-exchange DCBX with the peer device.  This
  generates events which trigger fcoemon to destroy the FCoE interface and
  then re-create it moments later due to the dcbd features momentarily going
  operationally false and then true again.

This patch modifies the state machine of fcoemon such that it is not
highly reactive to any single event or change.  Events (link and dcbd) are
handled on a continuous basis (instead of being filtered by time).  In any
given cycle of the fcoemon core loop, all events are processed first and
just update the state of the elements in the FCoE and network interface
lists.  At the end, the results are evaluated and any actions that need to
occur are triggered.

If the result is a completed dcbd query sequence, then the evaluation of the
results follow a new policy which requires specific criteria
to be satisfied before a create or destory action is taken on an FCoE
interface.  If the specific criteria is not met, then no change is made to
the FCoE interface.  The result is such that once an interface is created,
intermittent changes to the dcbd configuration will not cause the interface
to be destroyed.

Following are the required conditions to create an FCoE interface:
1.  Link is up on the network interfaces required for the FCoE interface, and

2a. DCB is not required.
	OR
2b. DCB is required
	AND DCB is enabled
	AND PFC is enabled
	AND App:FCoE is enabled
	AND PFC and App:FCoE are operational
	AND PFC and App:FCoE operational values match

Following are the required conditions to destroy an FCoE interface:
1.  The network interface required for the FCoE interface is removed
	Such as driver is unloaded or VLAN interface is destroyed.
	This is detected by DELLINK rtnetlink events.

2.  The network interfaces required for the FCoE interface are up,
	AND DCB is required,
	AND on root network interface
		DCB is disabled
			OR
		App:FCoE is disabled
			OR
		PFC and App:FCoE are operational
		AND PFC and App:FCoE operational values mismatch

If the above conditions for create or destroy are not met after handling
all incoming link and dcbd events, then no change is made to the FCoE
interface.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>