Add dcbd query timeout and retry mechanism
Previously, fcoemon would tend to destroy the interface too quickly, often
because it was essentially racing with dcbd.  fcoemon could finish its dcbd
queries and determine that dcbd state was bad before dcbd had completed its
synchronization process with the peer.

A dcbd query timeout and retry mechanism is added.  Anytime the dcbd query
sequence has an error or fails to complete with a definitive create or destroy
FCoE interface result, the query state machine will keep retrying up to a
limit.  If, at the end, the dcbd query sequence still does not result in a
create interface result, only then is the FCoE interface is destroyed.
The network interface must be up for this to occur.  Members are added to the
FCoE and network interface structures to maintain the current state of the
network interface. dcbd queries are not attempted for deleted or down network
interfaces.

This mechansism is provided as a way to destroy the FCoE interface when
dcb is required and in the atypical cases where a unusual, persistent dcb
configuration change has occurred.  (All typical use cases will just
expericence brief intermittent event sequences before settling down and
matching either the create or destroy criteria defined in the previous
patch.)

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>