fcoe-utils: make fcoemon and fcoeadm share a common version string
Report the common package version for all the tools, in an attempt to be less
confusing.  This makes fcoeadm, fcoemon and fipvlan all report 1.0.8 now.

In adding the new autoconf generated fcoe_utils.h header file to .gitignore, I
noticed that the resultant files for the passthrough tools (fcping, fcnsq,
fcrls) were not listed there either, so I tacked that on here as well.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>