fcoe-utils: Create contrib/ directory and move fcc.sh into it
fcc.sh isn't really a debug utility, it's more of a user contribution.
Also, there will be a FCoE EDD reading script that should be published
soon. It doesn't make sense to create a boot/ directory for this one
EDD script.

This patch creates a contrib/ directory and moves fcc.sh into it. This
contrib directory is also where the EDD script will live.

Signed-off-by: Robert Love <robert.w.love@intel.com>