fcoe-utils: Make FCOE_ENABLED="yes" the default
Make the entry in cfg-ethx for FCOE_ENABLED="yes" by
default. This allows a user doing a source install to
simply copy the template without having to edit the
resultant copy.

The only problem I can see with this change is if distros
create a cfg-eth[0-9] file for each interface on install
because then you'd have fcoe enabled for every interface
in the system. However, this is not the case, it is up to
the user to create the config files after fcoe-utils is
installed. This change might need to be reverted in the
future if the distros change that behavior.

Signed-off-by: Robert Love <robert.w.love@intel.com>