fcoe-utils: Remove initd.fcoe from dist_noinst_SCRIPTS and fix Makefile syntax
Even though initd.fcoe has been removed from the repository
it's still listed in the Makefile. This causes fcoe-utils
to fail to build.

This patch removes the initd.fcoe entry from dist_noinst_SCRIPTS
so that the build doesn't expect a rule for that non-existent
file.

There's also a syntax error in Makefile.am when the initd script
is installed. This patch fixes that error by making the SuSE
initd script installed by default if a Fedora or RedHat release
file isn't found.

Signed-off-by: Robert Love <robert.w.love@intel.com>