fcoemon: Add support for enable/disable of the FCoE interface in fcoemon.
The fcoe stack now supports enable and disable of an FCoE interface in addition
to create and destroy.  This patch updates fcoemon to take advantage of
enable and disable such that, once an FCoE inteface has been created, it is
only destroyed by a user administrative action (fcoeadm), the network driver
is unloaded or when fcoemon terminates.  Other conditions where fcoemon
previously destroyed the interface (e.g. misconfigured DCB parameters) now
result in the interface being disabled instead.

As part of this change, the logic for setting the next action to take place
on an FCoE interface (i.e. create, destroy, enable, etc.) is consolidated
into a new routine fcm_set_next_action() - which takes into account the
last action which occurred on the interface when determining the next action.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>