fcoemon: remove the qos_mask from the fcm_netif structure
The qos_mask was used to track the priority used for FCoE and whether it
changed or not.  This information was used to decide whether to create
or reset the FCoE interface when it was found to have good DCB parameters.
Also, the fcoeplumb script was invoked and that adjusted the tc filters for
the new priority.  Now that the tc filters are no longer used, it is not
necessary for fcoemon to keep track of the FCoE priority - so the
qos_mask can just be removed.

Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>