fcoemon: Use strncpy not memcpy when copying sun_path
Occasionally fcoemon fails to reply to a fcoeadm request.
When this happens fcoeadm eventually times out. It is the
result of using memcpy and not strncpy when copying the
sun_path received from the request.

Using memcpy is causing intermittent problems when copying
the sun_path from socket information. sun_path is a string
and therefore strncpy should be used. memcpy is not going to
copy the terminating character becuase we're using strlen
to determine the number of bytes to be copied.

The problem is masked becuase much of the time the destination
buffer must have been allocated from NULL'd out memory, but if
the allocated buffer started with junk characters then the code
ends up copying into a buffer without termination. So when we
try to reply to fcoeadm the sendto interface complains that
the sun_path is invalid (becuase there is garbage in the string).

This patch makes it so strncpy is used since we're dealing
with strings. It also changes the last argument to be the
size of the destination buffer as that's what strncpy requires.

Signed-off-by: Robert Love <robert.w.love@intel.com>