fcoe-utils: Update QUICKSTART guide
This patch adds more details to the QUICKSTART guide so that
it includes kernel configuration and user space configuration.

I've also changed some formatting. The column width is now 80
characters. The only problem with this is that the DCB link
had to be split in half.

I cleaned up some grammatical errors and tried to add more
helpful information when applicable.

Signed-off-by: Robert Love <robert.w.love@intel.com>