fcoeadm: Simplify output of -v (version) option
The copyright doesn't have anything to do with the version. This
patch removes that output from the 'fcoeadm -v' command. It also
removes the preceeding 'v' from the version. So, instead of

# fcoeadm -v
fcoeadm v1.0.10
 Copyright (c) 2009, Intel Corporation.

it will instead print

# fcoeadm -v
1.0.10

This change also alters the --help option as it was printing
the version at the top of its output. Now it just shows-

]# fcoeadm --help
Version 1.0.10
Usage: fcoeadm
        [-c|--create] <ethX>
	[-d|--destroy] <ethX>
	[-r|--reset] <ethX>
	[-i|--interface] [<ethX>]
	[-t|--target] [<ethX>]
	[-l|--lun] [<target port_id> [<lun_id>]]
	[-s|--stats] <ethX> [-n <interval>]
	[-v|--version]
	[-h|--help]

Signed-off-by: Robert Love <robert.w.love@intel.com>