fcoeadm: Create a common routine to find interface from symbolic_name
There are a few places that do this operation with strstr, but
this is broken because if the user specifies "eth3" and there's
only a connection on "eth3.101" it will incorrectly match.

This patch makes one routine that can be used in multiple places
and will have a stronger way to ensure the user specified data
is correct.

While making this change I noticed another problem that is fixed
in this patch. The ifname member of opt_info is a char[16] which
means that the 'ifname' member is never NULL. The code is often
checking if 'ifname' is NULL to indicate that no ifname was passed
in by the user. Since 'ifname' will never be NULL then the intent
of the checks is never achieved. This patch creates a valid_ifname()
routine in the new fcoe_utils.[ch] that checks the string length
to determine if the ifname is valid.

Signed-off-by: Robert Love <robert.w.love@intel.com>