fcoeadm: Don't pass ifname into _cli_open
This code is strange. It scans the /var/run/fcm/
directory for anything and if it finds something other
than "." or ".." it assumes that it's the socket
file.

Since the socket file is agreed upon through fcoe_clif.h
there's no need to try and find it. Let's just use
the agreed upon file and not do the scanning.

Signed-off-by: Robert Love <robert.w.love@intel.com>