fcoeadm: Improve getopt_long case statement
This patch improves the main() fcoeadm routine. It
attempts to make the getopt_long processing more
readable and adds a single error handling switch-
statement.

Going forward we should not be printing error
messages throughout the code and should allow
the common error handler to print error messages
so that they're common.

This patch also removes the optional '-n' flag option
from the '-s' option. The usage changes from-

fcoeadm -s ethX.101 [ -n <interval> ]

to

fcoeadm -s ethX.101 [ <interval> ]

Finally, this patch limits the behavior of the '-l'
option. Now users cannot isolate a single lun. The
only option available is the interface name and the
'-l' option will print all luns on that interface.

The problem with '-l' is that LUNs are enumerated on
the target by the connection to the target, so to
specify a single LUN we would need to provide,
interface name, target ID and LUN ID. It isn't clear
that users value isolating a single LUN (it can still
be scripted) so rather than complicate things this
patch simplifies the option.

The usage changes from

fcoeadm [-l|--lun] [<target_port_id> [<lun_id>]]

to

fcoeadm [-l|--lun] [<ethX>]

The man page for fcoeadm is also updated to reflect
these changes. While updating the man page I found
some incorrect pluralization that I fixed as well.

Signed-off-by: Robert Love <robert.w.love@intel.com>